+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added `-NoWait` switch to `Copy-PnPFile` and `Move-PnPFile`
 - Added `Receive-PnPCopyMoveJobStatus` cmdlet which works in combination with the `-NoWait` parameter on `Copy-PnPFile` and `Move-PnPFile`. See the documentation for usage.
 
+### Changed
+- Fixed issue with using `-UseWebLogin` and site names longer than the length of the managed path it resides in.
 
 ## [1.1.2-nightly]